Move Ontology in AceWiki to another servlet.
Create a new servlet which hold the Ontology of a AceWiki, as the backend of
the AceWiki. The ontology object will share to AceWiki servlet by using
ServletContext. A change to web.xml is also necessary to make new AceWiki
work. First, create a servlet of class BackendServlet, move all original
parameters of AceWiki to the BackendServlet, and add a parameter "backend" to
AceWiki servlet to indicate which BackendServlet to use. See new web.xml
for more details.