Add "GF Module pages"
Very preliminary: Every GF module is currently stored in the first comment
of a special type of page ("GF module page"). When edited, the
content is pushed to the GF webservice which immediately recompiles
the module and updates the PGF. This way the grammar becomes
a proper part of the wiki and can be collaboratively edited.