Now storing: language label, original text, trees
If the stored tree cannot be linearized then the original text is
parsed to obtain a new tree. This makes the wiki insensitive
to tree changes (which can happen e.g. as a result of refactoring).

Also: remove "add sentence by random generation"

Also: Wiki page displays original input not the first variant,
i.e. "does not" is not replaced by "doesn't".