Support for copying article content; general 'advanced_commands' option
Statements can be copied from one article to another via a menu item in
the final gray triangle. For this, the parameter 'advanced_commands' has
to be set to 'on' and the user needs one of the rights
'advanced_commands' or 'all'. This includes the 'shuffle statements'
feature, which has no longer its own parameter or user right.